Function rust_audio_translate::s3::download_file

pub async fn download_file(
    client: &Client,
    bucket: String,
    key: String,
    output_path: &str,
) -> Result<(), Box<dyn Error>>

Downloads a file from an S3 bucket and saves it locally.

This function retrieves an object from the specified S3 bucket and saves it to the local filesystem.

Arguments

client - A reference to an S3Client used to interact with Amazon S3.

bucket - The name of the S3 bucket from which to download the file.

key - The object key (path) of the file in the S3 bucket.

output_path - The local file system path where the downloaded file will be saved.

Returns

Returns Ok(()) if the download and save operations are successful, or an error wrapped in a Box<dyn std::error::Error> if any step of the process fails.

Errors

This function will return an error if:

The S3 client fails to retrieve the object

The local file cannot be created or written to

Any other I/O or AWS SDK error occurs